+++ lib/tool_dispatch.ml
@@ -153,15 +153,15 @@ let is_join_required name = with_dispatch_ro (fun () -> Hashtbl.mem requires_joi
 type module_tag =
   | Mod_plan | Mod_run | Mod_operator | Mod_command_plane
   | Mod_local_runtime | Mod_team_session | Mod_voice | Mod_cache
-  | Mod_tempo | Mod_portal | Mod_worktree
+  | Mod_portal | Mod_worktree
   | Mod_code_swarm | Mod_code | Mod_code_write
   | Mod_council | Mod_a2a | Mod_handover
-  | Mod_relay | Mod_goals | Mod_heartbeat | Mod_encryption
+  | Mod_relay | Mod_goals | Mod_heartbeat
   | Mod_auth | Mod_hat | Mod_audit | Mod_rate_limit
   | Mod_cost | Mod_walph | Mod_agent | Mod_task | Mod_room
   | Mod_control | Mod_agent_timeline | Mod_misc | Mod_suspend
   | Mod_library | Mod_keeper | Mod_compact | Mod_mdal
-  | Mod_notifications | Mod_inline
+  | Mod_inline
   | Mod_improve_loop
   | Mod_autoresearch
   | Mod_research