+++ test/test_mcp_tool_matrix_cases.ml
@@ -131,9 +131,6 @@ let all_known_tool_names =
     "masc_dispatch_rebalance";
     "masc_dispatch_recall";
     "masc_dispatch_tick";
-    "masc_encryption_disable";
-    "masc_encryption_enable";
-    "masc_encryption_status";
     "masc_episode_flush";
     "masc_episode_list";
     "masc_error_add";
@@ -144,7 +141,6 @@ let all_known_tool_names =
     "masc_find_by_capability";
     "masc_fire_task";
     "masc_gc";
-    "masc_generate_key";
     "masc_get_metrics";
     "masc_goal_dispatch";
     "masc_goal_list";
@@ -316,11 +312,6 @@ let all_known_tool_names =
     "masc_team_session_step";
     "masc_team_session_stop";
     "masc_team_session_verify_trace";
-    "masc_tempo";
-    "masc_tempo_adjust";
-    "masc_tempo_get";
-    "masc_tempo_reset";
-    "masc_tempo_set";
     "masc_tool_admin_snapshot";
     "masc_tool_admin_update";
     "masc_tool_help";
@@ -346,7 +337,6 @@ let all_known_tool_names =
     "masc_voice_session_start";
     "masc_voice_sessions";
     "masc_voice_speak";
-    "masc_voice_transcript";
     "masc_walph_control";
     "masc_walph_natural";
     "masc_walph_status";
@@ -405,10 +395,6 @@ let strict_success_names =
     "masc_set_room";
     "masc_start";
     "masc_status";
-    "masc_tempo_adjust";
-    "masc_tempo_get";
-    "masc_tempo_reset";
-    "masc_tempo_set";
     "masc_tool_help";
     "masc_transition";
     "masc_transport_status";